+++ CHANGELOG.rst
@@ -3,7 +3,11 @@ CHANGELOG
 
 Release NEXT
 ------------
-- Nothing yet
+- Fix token authentication according to DRF 3.3.
+- Replace IPAddressField serializer with DRF serializer.
+- Add get_extra_kwargs method to AugmentedSerializerMixin.
+- Migrate to DRF 3.5.3 and django_filter 0.15.3.
+- Drop DjangoMappingFilterBackend and UUIDFilter.
 
 Release 0.118.0
 ---------------